zvkned: add vaesef.vv & vaesef.vs
The "vaesef.vv|vs" instruction performs the final-round
encryption of the AES block cipher.

The "SubBytes" and "ShiftRows" steps are applied to each round state
element group from "vd". This is then XORed with the corresponding round
key element group in "vs2"(vector-vector form) or the scalar element group
in "vs2"(vector-scalar form).

Signed-off-by: Charalampos Mitrodimas <charalampos.mitrodimas@vrull.eu>