zvkned: add vaeskf2.vi
The "vaeskf2.vi" instruction performs a single round
of the forward AES-256 KeySchedule generation.

The next round key is generated word by word from the previous round
key element group in "vd" and the immediately previous word of the round
key. The least significant word of the next round key is generated by
applying a function to the most significant word of the current round
key and then XORing the result with the round constant. The round
number is used to select the round constant as well as the function.

The round number(ranges 2-14), comes from imm[3:0]; imm[4] is ignored.
We project out-of-range immediates into in-range values by inverting
imm[3].

Signed-off-by: Charalampos Mitrodimas <charalampos.mitrodimas@vrull.eu>